requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mistralai==1.0.0
python-dotenv==1.0.0
pydantic>=2.8.2,<2.9.0
